server/requirements.txt
backcall==0.1.0
CacheControl==0.12.6
cachetools==4.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
decorator==4.4.2
firebase-admin==4.0.1
Flask==1.1.1
Flask-Cors==3.0.8
gcloud==0.17.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
httplib2==0.17.1
idna==2.9
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jws==0.1.3
MarkupSafe==1.1.1
msgpack==1.0.0
oauth2client==3.0.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pygments==2.6.1
Pyrebase==3.0.27
python-dotenv==0.12.0
python-jwt==2.0.1
pytz==2019.3
requests==2.23.0
requests-toolbelt==0.7.0
rsa==4.0
six==1.14.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.0
